💥 enable network errors removal (#938)
* 💥 Remove network error tracking from RUM

* ♻️ Move trackNetworkError from core to logs

* Remove remove-network-errors flag
