+++ packages/core/src/index.ts
@@ -7,7 +7,6 @@ export {
   BeforeSendCallback,
 } from './domain/configuration'
 export { trackConsoleError } from './domain/error/trackConsoleError'
-export { trackNetworkError } from './domain/error/trackNetworkError'
 export { trackRuntimeError } from './domain/error/trackRuntimeError'
 export { computeStackTrace, StackTrace } from './domain/tracekit'
 export {
