+++ packages/logs/src/domain/trackNetworkError.ts
@@ -1,24 +1,23 @@
-import { FetchCompleteContext, resetFetchProxy, startFetchProxy } from '../../browser/fetchProxy'
-import { resetXhrProxy, startXhrProxy, XhrCompleteContext } from '../../browser/xhrProxy'
-import { ErrorSource, RawError } from '../../tools/error'
-import { Observable } from '../../tools/observable'
-import { RequestType } from '../../tools/utils'
-import { Configuration } from '../configuration'
+import {
+  Configuration,
+  ErrorSource,
+  FetchCompleteContext,
+  Observable,
+  RawError,
+  RequestType,
+  resetFetchProxy,
+  resetXhrProxy,
+  startFetchProxy,
+  startXhrProxy,
+  XhrCompleteContext,
+} from '@datadog/browser-core'
 
-export function trackNetworkError(
-  configuration: Configuration,
-  errorObservable: Observable<RawError>,
-  trackAbortedRequests = true
-) {
+export function trackNetworkError(configuration: Configuration, errorObservable: Observable<RawError>) {
   startXhrProxy().onRequestComplete((context) => handleCompleteRequest(RequestType.XHR, context))
   startFetchProxy().onRequestComplete((context) => handleCompleteRequest(RequestType.FETCH, context))
 
   function handleCompleteRequest(type: RequestType, request: XhrCompleteContext | FetchCompleteContext) {
-    if (
-      !configuration.isIntakeUrl(request.url) &&
-      (trackAbortedRequests || !request.isAborted) &&
-      (isRejected(request) || isServerError(request))
-    ) {
+    if (!configuration.isIntakeUrl(request.url) && (isRejected(request) || isServerError(request))) {
       errorObservable.notify({
         message: `${format(type)} error ${request.method} ${request.url}`,
         resource: {
