+++ packages/rum-core/src/boot/startRum.ts
@@ -58,7 +58,7 @@ export function startRum(
     foregroundContexts,
     initialViewName
   )
-  const { addError } = startErrorCollection(lifeCycle, configuration, foregroundContexts)
+  const { addError } = startErrorCollection(lifeCycle, foregroundContexts)
   const { addAction } = startActionCollection(lifeCycle, domMutationObservable, configuration, foregroundContexts)
 
   startRequestCollection(lifeCycle, configuration)
